Move template/interface/choice processing into TcDaml
