+++ compiler/typecheck/TcDaml.hs
@@ -1,6 +1,4 @@
-module TcDaml(
-        customDamlErrors
-  ) where
+module TcDaml where
 
 import GhcPrelude
 
@@ -13,6 +11,16 @@ import OccName
 import Name
 import Module
 import PrelNames
+import RdrHsSyn
+import RdrName
+import SrcLoc
+import HsTypes
+import HsDecls
+import HsExtension
+import Type
+
+import Data.Maybe
+import Data.List
 
 check :: [String] -> String -> TyCon -> Bool
 check modules name con
@@ -44,3 +52,73 @@ customDamlErrors ct
   $ text "Tried to implement method" <+> ppr methodName <> text ", but interface" <+> ppr (tyConName target) <+> text "does not have a method with that name."
   | otherwise
   = pure Nothing
+
+data DamlVariant = Template Name | Interface Name | Choice Name
+
+data TyConDamlVariant = TyConTemplate Name AlgTyConRhs | TyConInterface Name AlgTyConRhs | TyConChoice Name String
+
+instance Outputable TyConDamlVariant where
+  ppr (TyConTemplate name _) = text "TyConTemplate" <+> ppr name
+  ppr (TyConInterface name _) = text "TyConInterface" <+> ppr name
+  ppr (TyConChoice target name) = text "TyConChoice" <+> ppr target <> text name
+
+tyConDamlVariant_maybe :: TyCon -> Maybe TyConDamlVariant
+tyConDamlVariant_maybe tycon
+  | isAlgTyCon tycon
+  , any (isConstraint ghcTypesDamlInterface) (tyConStupidTheta tycon)
+  = Just $ TyConInterface (tyConName tycon) (algTyConRhs tycon)
+  | isAlgTyCon tycon
+  , any (isConstraint ghcTypesDamlTemplate) (tyConStupidTheta tycon)
+  = Just $ TyConTemplate (tyConName tycon) (algTyConRhs tycon)
+  | otherwise
+  = Nothing
+  where
+    isConstraint :: RdrName -> Type -> Bool
+    isConstraint rdrName type_
+      | Just (loneConstraint, []) <- splitTyConApp_maybe type_
+      , let loneConstraintName = tyConName loneConstraint
+      , similarName rdrName loneConstraintName
+      = True
+    isConstraint _ _ = False
+
+addDamlTypesToGblEnv :: [LTyClDecl GhcRn] -> TcGblEnv -> TcGblEnv
+addDamlTypesToGblEnv tyClDecls env@(TcGblEnv { tcg_daml_templates = templates
+                                             , tcg_daml_interfaces = interfaces
+                                             , tcg_daml_choices = choices }) =
+  let (newTemplates, newInterfaces, newChoices) = extractDamlTypes tyClDecls
+  in
+  env { tcg_daml_templates  = templates ++ newTemplates
+      , tcg_daml_interfaces = interfaces ++ newInterfaces
+      , tcg_daml_choices    = choices ++ newChoices
+      }
+
+extractDamlTypes :: [LTyClDecl GhcRn] -> ([Name], [Name], [Name])
+extractDamlTypes = splitVariants . mapMaybe extractDamlType
+  where
+    splitVariants :: [DamlVariant] -> ([Name], [Name], [Name])
+    splitVariants = foldMap $ \x -> case x of
+                                     Template a -> ([a], mempty, mempty)
+                                     Interface a -> (mempty, [a], mempty)
+                                     Choice a -> (mempty, mempty, [a])
+
+extractDamlType :: LTyClDecl GhcRn -> Maybe DamlVariant
+extractDamlType (L _ DataDecl { tcdLName = L _ name, tcdDataDefn = HsDataDefn { dd_ctxt = L _ contextTypes } })
+  | any (isConstraint ghcTypesDamlInterface) contextTypes
+  = Just (Interface name)
+  | any (isConstraint ghcTypesDamlTemplate) contextTypes
+  = Just (Template name)
+  where
+    isConstraint :: RdrName -> LHsType GhcRn -> Bool
+    isConstraint rdrName (L _ (HsTyVar _ _ (L _ loneConstraint))) =
+      similarName rdrName loneConstraint
+    isConstraint _ _ = False
+extractDamlType _ = Nothing
+
+similarName :: RdrName -> Name -> Bool
+similarName (Qual targetModuleName targetOccName) name
+  | Just actualModule <- nameModule_maybe name
+  , moduleName actualModule == targetModuleName
+  , nameOccName name == targetOccName
+  = True
+similarName _ _ = False
+
